requirements.txt
django
djangorestframework
dotenv
mysqlclient
langchain_google_genai
gradio_client